refactor: extract scan pipeline from api/server.py into api/pipeline.py (Phase 4)
Move ScanPipeline class, _run_scan_sync, _sync_scan_agents_to_fleet,
_now helper, and PIPELINE_STEPS constant into dedicated pipeline module.
server.py drops from 3,558 to 3,127 lines (-431).

Total api/server.py reduction: 4,154 → 3,127 across all 4 phases:
- Phase 1+2: models.py (258 lines) + middleware.py (272 lines)
- Phase 3: stores.py (207 lines)
- Phase 4: pipeline.py (465 lines)

All symbols re-exported from server.py for backward compatibility.
Patch targets in test_medium_hardening.py updated to pipeline module.
223 API tests pass, zero regressions.
